edited ProductionManager

+++ src/main/java/com/example/infraboxapi/productionItem/ProductionItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductionItemRepository extends JpaRepository<ProductionItem, Long> {
+public interface ProductionItemRepository extends JpaRepository<ProductionItem, Integer> {
 }
